Generate PDBs with new PAMs given PDB with Cas9 variant
This script will, given a PDB containing the Cas9 variant of interest,
generate new PDBs with the desired PAM site.

It is set up for PAMs of length 3, and the only thing needed to take in
length 4 is to find the right position in the two chains to modify.

This is adapted from a script @mglubber wrote