SetPyRosettaEnvironment.sh that allows you to use PyRosetta outside of the source folder, and PyMOL Observer script. Have a PyMOL instance open and listening for PyRosetta before running.
Currently pyobs updates on ANY event (change to pose); ~1400 events occured with DockMCMProtocol before we stopped it, and then PyMOL killed the comp with RAM usage trying to visualize those results. So visualize a smaller protocol, or might be able to use pyobs.generalEvent(parameter) in some way to change when it PyMOL updates.