Refactored viral assembly code
Split code to two files, one with the ODE and the other runs the ODE
and plots it. The ODE has a few changes, notably P4 subunits removed,
P2->elIB term removed, outside infection removed, and some changes to
make it look neater. Parameter values are no longer contained in the
ODE file, they’re passed to it when it’s called.